requirements.txt
streamlit
pandas
psycopg2-binary
python-dotenv
streamlit-cookies-manager
python-frontmatter
markdown